Remove use of RssEntry in Expansion.java
Remove the import of RssEntry and the check made in equals() method
between the expansion name and the RssEntry title.
This was done to make Expansion.java fully independant from other
classes of the project. This was donne to facilitate its export to the
mtg-familiar project, to share the core classes used in a patch, so that
gson car be used in both project, using the sames classes and code.